Separate helper functions to helper.js file
